Converting AWS KMS filter plugin to lookup plugins
